Include itk-elastix in tomviz packaging
We want to include `itk-elastix` as a dependency to tomviz, but it is
not available on conda-forge.

However, it is available on pip, it is small, and it was designed to
work in conda environments (similar to ITK). Thus, we can `pip install`
it and include it in our tomviz package.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>